+++ include/nnue.h
@@ -9,46 +9,8 @@
 
 #include "constants.h"
 #include "bitboard.h"
+#include "simd.h"
 #include "weights.h"
-
-// activation is implicitly ReLU, except for L2 -> Out which is linear.
-
-const __m256i _ZERO = _mm256_setzero_si256();
-const __m256i _ONE = _mm256_set1_epi16(1);
-
-// horizontal sum of vector
-// https://stackoverflow.com/questions/60108658/fastest-method-to-calculate-sum-of-all-packed-32-bit-integers-using-avx512-or-av
-
-inline int hsum_epi32_avx(__m128i x)
-{
-    __m128i hi64 = _mm_unpackhi_epi64(x, x);
-    __m128i sum64 = _mm_add_epi32(hi64, x);
-    __m128i hi32 = _mm_shuffle_epi32(sum64, _MM_SHUFFLE(2, 3, 0, 1));
-    __m128i sum32 = _mm_add_epi32(sum64, hi32);
-    return _mm_cvtsi128_si32(sum32);
-}
-
-inline int hsum_8x32(__m256i v)
-{
-    __m128i sum128 = _mm_add_epi32(
-        _mm256_castsi256_si128(v),
-        _mm256_extracti128_si256(v, 1));
-    return hsum_epi32_avx(sum128);
-}
-
-inline __m256i cvtepi16_epi8(__m256i low, __m256i high)
-{
-    __m256i res = _mm256_packs_epi16(low, high);
-    res = _mm256_permute4x64_epi64(res, 0b11011000);
-    return res;
-}
-
-inline __m256i madd_epi8(__m256i a, __m256i b)
-{
-    __m256i res = _mm256_maddubs_epi16(a, b);
-    res = _mm256_madd_epi16(res, _ONE);
-    return res;
-}
 
 class Accumulator
 {
@@ -75,7 +37,7 @@ public:
         _move(move, &Accumulator::setOne, &Accumulator::setZero);
     }
 
-    void _move(U32 move, void (Accumulator::*_zero)(U32 index), void (Accumulator::*_one)(U32 index))
+    void _move(U32 move, void (Accumulator::*_zero)(U32 x, U32 y), void (Accumulator::*_one)(U32 x, U32 y))
     {
         U32 pieceType = (move & MOVEINFO_PIECETYPE_MASK) >> MOVEINFO_PIECETYPE_OFFSET;
         if (pieceType == side)
@@ -90,11 +52,11 @@ public:
         U32 capturedPieceType = (move & MOVEINFO_CAPTUREDPIECETYPE_MASK) >> MOVEINFO_CAPTUREDPIECETYPE_OFFSET;
         bool enPassant = move & MOVEINFO_ENPASSANT_MASK;
 
-        (this->*_zero)(index(pieceType, startSquare));
-        (this->*_one)(index(finishPieceType, finishSquare));
+        (this->*_zero)(pieceType, startSquare);
+        (this->*_one)(finishPieceType, finishSquare);
         if (capturedPieceType != 15)
         {
-            (this->*_zero)(index(capturedPieceType, finishSquare + enPassant * (-8 + 16 * (pieceType & 1))));
+            (this->*_zero)(capturedPieceType, finishSquare + enPassant * (-8 + 16 * (pieceType & 1)));
         }
 
         // check for castles.
@@ -102,13 +64,13 @@ public:
         {
             if (finishSquare - startSquare == 2)
             {
-                (this->*_zero)(index(_nRooks + !side, KING_ROOK_SQUARE[!side]));
-                (this->*_one)(index(_nRooks + !side, KING_ROOK_SQUARE[!side] - 2));
+                (this->*_zero)(_nRooks + !side, KING_ROOK_SQUARE[!side]);
+                (this->*_one)(_nRooks + !side, KING_ROOK_SQUARE[!side] - 2);
             }
             else if (startSquare - finishSquare == 2)
             {
-                (this->*_zero)(index(_nRooks + !side, QUEEN_ROOK_SQUARE[!side]));
-                (this->*_one)(index(_nRooks + !side, QUEEN_ROOK_SQUARE[!side] + 3));
+                (this->*_zero)(_nRooks + !side, QUEEN_ROOK_SQUARE[!side]);
+                (this->*_one)(_nRooks + !side, QUEEN_ROOK_SQUARE[!side] + 3);
             }
         }
     }
@@ -118,32 +80,34 @@ public:
         kingPos = __builtin_ctzll(pieces[side]);
         std::copy(b_0.begin(), b_0.end(), l1);
 
-        setOne(index(!side, __builtin_ctzll(pieces[!side])));
+        setOne(!side, __builtin_ctzll(pieces[!side]));
         for (size_t i = 2; i < 12; ++i)
         {
             U64 x = pieces[i];
             while (x)
             {
-                setOne(index(i, popLSB(x)));
+                setOne(i, popLSB(x));
             }
         }
         cReLU();
     }
 
-    void setOne(U32 index)
+    void setOne(U32 pieceType, U32 square)
     {
+        U32 ind = index(pieceType, square);
         for (size_t i = 0; i < 4; ++i)
         {
-            l1[i] = _mm256_add_epi16(l1[i], w_0[index][i]);
+            l1[i] = _mm256_add_epi16(l1[i], w_0[ind][i]);
         }
         cReLU();
     }
 
-    void setZero(U32 index)
+    void setZero(U32 pieceType, U32 square)
     {
+        U32 ind = index(pieceType, square);
         for (size_t i = 0; i < 4; ++i)
         {
-            l1[i] = _mm256_sub_epi16(l1[i], w_0[index][i]);
+            l1[i] = _mm256_sub_epi16(l1[i], w_0[ind][i]);
         }
         cReLU();
     }
